refactor(game): cleanup
