+++ src/game/reducers/README.md
@@ -1,6 +1,3 @@
 # Game Reducers
 
-The functions exported from this directory are meant to implement Farmhand
-Shuffle game logic. These functions **must** be
-[pure](https://www.freecodecamp.org/news/what-is-a-pure-function-in-javascript-acb887375dfe/)
-and treat data [immutably](https://en.wikipedia.org/wiki/Immutable_object).
+The functions exported from this directory are meant to implement Farmhand Shuffle game logic. These functions **must** be [pure](https://www.freecodecamp.org/news/what-is-a-pure-function-in-javascript-acb887375dfe/) and treat data [immutably](https://en.wikipedia.org/wiki/Immutable_object).
