+++ src/game/reducers/update-deck/index.test.ts
@@ -1,5 +1,5 @@
 import { stubGame } from '../../../test-utils/stubs/game'
-import { carrot } from '../../cards/crops/carrot/carrot'
+import { carrot } from '../../cards/crops/carrot'
 
 import { updateDeck } from './'
