+++ src/game/services/README.md
@@ -1,3 +1,5 @@
 # Game Services
 
-This code in this directory is concerned with Farmhand Shuffle game logic and behavior.
+This code in this directory is concerned with managing and executing Farmhand Shuffle game logic. Services are broader in scope than reducers, which are solely concerned with transforming data.
+
+The service pattern is inspired by [Angular services](https://angular.io/guide/architecture-services).
